[Build] Complete the loadlib relocation out of dep/
The earlier move of the loadlib reader out of dep/ landed the file move
but not the build wiring, leaving the tree unable to configure or compile
the extractors (the loadlib target was never defined and stale include
paths pointed at the removed dep/loadlib/sl/ location).

- dep/CMakeLists.txt: drop the dangling add_subdirectory(loadlib) (the
  directory no longer lives under dep/)
- src/tools/Extractor_projects/CMakeLists.txt: add_subdirectory(loadlib)
  so the relocated target is actually built
- fix four stale includes ../loadlib/sl/X.h -> ../loadlib/X.h
  (map-extractor System.cpp x2, dbcfile.cpp, shared/ExtractorCommon.h)
- update the now-stale dep/loadlib/sl/adt.h reference in a GridMap comment

map-extractor, vmap-extractor and mmap-extractor all build again.
